+++ Changes.md
@@ -28,6 +28,11 @@ Build
 
 - Cortex : Updated to version 10.5.13.0.
 
+API
+---
+
+- Attributes, Options : Added protected constructors for initialising from attributes/options defined by metadata.
+
 1.5.6.0 (relative to 1.5.5.0)
 =======